ADC: updated to HAL G0

+++ src/adc.h
@@ -33,8 +33,8 @@
 typedef struct builtinAdcSetup{
 	GPIO_TypeDef	*gpio;
 	uint16_t		pin;
-	uint8_t			channel;
-	uint8_t			sample_time; //must be ADC_SAMPLINGTIME_COMMON_1 or ADC_SAMPLINGTIME_COMMON_2
+	uint32_t		channel;
+	uint32_t		sample_time; //must be ADC_SAMPLINGTIME_COMMON_1 or ADC_SAMPLINGTIME_COMMON_2
 } builtinAdcSetup;
 
 void ADC_Init(uint16_t *adc_buffer, uint32_t num_channels, builtinAdcSetup *adc_setup);
